backend/app/api/v1/endpoints/graph.py
"""
Graph endpoints.

Graph traversal and visualization endpoints.
"""

from fastapi import APIRouter, Depends
from arango.database import StandardDatabase

from app.db.arangodb import get_arango_db

router = APIRouter()


@router.get("/")
async def get_graph_data(
    db: StandardDatabase = Depends(get_arango_db)
):
    """
    Get graph data for visualization.

    Args:
        db: ArangoDB database instance

    Returns:
        dict: Graph nodes and edges for visualization
    """
    # TODO: Implement graph data retrieval

    return {
        "nodes": [],
        "edges": []
    }


@router.get("/{node_id}/neighborhood")
async def get_node_neighborhood(
    node_id: str,
    depth: int = 1,
    db: StandardDatabase = Depends(get_arango_db)
):
    """
    Get neighborhood of a graph node.

    Args:
        node_id: Node ID
        depth: Traversal depth (1-3)
        db: ArangoDB database instance

    Returns:
        dict: Nodes and edges in neighborhood
    """
    # TODO: Implement neighborhood traversal

    return {
        "nodes": [],
        "edges": []
    }
